refactor: 픽스쳐 추가 및 디렉토리 수정
