+++ melly-core/src/test/java/cmc/mellyserver/mellycore/MellyCoreApplicationTests.java
@@ -2,10 +2,9 @@ package cmc.mellyserver.mellycore;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.test.context.ActiveProfiles;
+
 
 @SpringBootApplication
-@ActiveProfiles("test")
 public class MellyCoreApplicationTests {
 
     @Test
