--- melly-core/src/test/java/cmc/mellyserver/mellycore/common/fixture/PlaceFixtures.java
@@ -1,21 +1,0 @@
-package cmc.mellyserver.mellycore.common.fixture;
-
-public class PlaceFixtures {
-//
-//    /* 스타벅스 */
-//    public static final String 스타벅스 = "스타벅스";
-//
-//    public static final Position 스타벅스_좌표 = new Position(1.234, 1.234);
-//    public static final String 스타벅스_이미지 = "/starbucks.png";
-//    public static final String 스타벅스_카테고리 = "카페/베이커리";
-//
-//    public static Place 스타벅스() {
-//        return Place.builder()
-//                .position(스타벅스_좌표)
-//                .placeName(스타벅스)
-//                .placeImage(스타벅스_이미지)
-//                .placeCategory(스타벅스_카테고리)
-//                .isDeleted(Boolean.FALSE)
-//                .build();
-//    }
-}
