+++ melly-core/src/test/java/cmc/mellyserver/mellycore/config/RepositoryTest.java
@@ -1,13 +1,12 @@
-package cmc.mellyserver.mellycore.common;
+package cmc.mellyserver.mellycore.config;
 
-import cmc.mellyserver.mellycore.config.JpaTestConfiguration;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 
 @DataJpaTest
 @ActiveProfiles("test")
-@Import(JpaTestConfiguration.class)
+@Import(JpaTestConfig.class)
 public abstract class RepositoryTest {
 
 }
