--- melly-core/src/test/java/cmc/mellyserver/mellycore/config/mockapi/InfrastructureTestConfiguration.java
@@ -1,12 +1,0 @@
-package cmc.mellyserver.mellycore.config.mockapi;
-
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-
-@Configuration
-@Profile("test")
-public class InfrastructureTestConfiguration {
-
-
-}
