--- melly-core/src/test/java/cmc/mellyserver/mellycore/user/application/UserSurveyServiceTest.java
@@ -1,4 +1,0 @@
-package cmc.mellyserver.mellycore.user.application;
-
-public class UserSurveyServiceTest {
-}
